game changed to MainMenu instead of MainPage
